pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-bench-cli"
version = "0.1.0"
description = "A developer-centric CLI tool to systematically evaluate and compare Large Language Models (LLMs)"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [
    { name = "Abdul Basit Banbhan" }
]
keywords = ["llm", "benchmark", "cli", "ai", "testing", "evaluation", "gpt", "claude", "gemini", "openai", "anthropic"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Quality Assurance",
    "Typing :: Typed",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "litellm>=1.0.0",
    "pydantic>=2.0.0",
    "deepdiff>=6.0.0",
    "diskcache>=5.6.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/abdulbb/llm-bench-cli"
Repository = "https://github.com/abdulbb/llm-bench-cli.git"
Issues = "https://github.com/abdulbb/llm-bench-cli/issues"
Documentation = "https://github.com/abdulbb/llm-bench-cli#readme"
Changelog = "https://github.com/abdulbb/llm-bench-cli/releases"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
llm-bench = "llm_bench.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_bench"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
addopts = [
    "--cov=llm_bench",
    "--cov-report=term-missing",
    "--cov-report=html",
    "-v",
]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = [
    "litellm.*",
    "diskcache.*",
    "deepdiff.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py310"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["llm_bench"]

[tool.coverage.run]
source = ["src/llm_bench"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]